+++ src/sll/api/file_api.c
@@ -6,13 +6,14 @@
 #include <sll/data.h>
 #include <sll/file.h>
 #include <sll/gc.h>
+#include <sll/init.h>
 #include <sll/memory.h>
 #include <sll/object.h>
 #include <sll/platform.h>
+#include <sll/sandbox.h>
 #include <sll/static_object.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>
 #include <sll/vm.h>