+++ src/sll/api/json_api.c
@@ -6,6 +6,7 @@
 #include <sll/common.h>
 #include <sll/data.h>
 #include <sll/file.h>
+#include <sll/init.h>
 #include <sll/map.h>
 #include <sll/memory.h>
 #include <sll/object.h>
@@ -13,7 +14,6 @@
 #include <sll/static_object.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>
 #include <sll/vm.h>
 #include <stdio.h>