+++ src/sll/api/path_api.c
@@ -9,10 +9,10 @@
 #include <sll/memory.h>
 #include <sll/object.h>
 #include <sll/platform.h>
+#include <sll/sandbox.h>
 #include <sll/static_object.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>