+++ src/sll/platform/windows/win_util.c
@@ -3,10 +3,10 @@
 #include <sll/_sll_internal.h>
 #include <sll/common.h>
 #include <sll/data.h>
+#include <sll/init.h>
 #include <sll/memory.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>
 #include <signal.h>