+++ src/sll/vm.c
@@ -10,6 +10,7 @@
 #include <sll/object.h>
 #include <sll/operator.h>
 #include <sll/platform.h>
+#include <sll/sandbox.h>
 #include <sll/static_object.h>
 #include <sll/types.h>
 #include <sll/util.h>