+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 ### Changed
 
 - Moved `SLL_COPY_STRING_NULL`, `sll_compare_data`, `sll_copy_data`, `sll_copy_string`, `sll_copy_string_null`, `sll_set_memory` and `sll_zero_memory` from [`src/sll/util.c`][0.6.37/src/sll/util.c] and [`src/sll/include/sll/util.h`][0.6.37/src/sll/include/sll/util.h] to [`src/sll/data.c`][0.6.38/src/sll/data.c] and [`src/sll/include/sll/data.h`][0.6.38/src/sll/include/sll/data.h]
+- Moved `sll_deinit`, `sll_init` and `sll_register_cleanup` from [`src/sll/util.c`][0.6.37/src/sll/util.c] and [`src/sll/include/sll/util.h`][0.6.37/src/sll/include/sll/util.h] to [`src/sll/init.c`][0.6.38/src/sll/init.c] and [`src/sll/include/sll/init.h`][0.6.38/src/sll/include/sll/init.h]
+- Moved `sll_get_environment_variable`, `sll_remove_environment_variable` and `sll_set_environment_variable` from [`src/sll/util.c`][0.6.37/src/sll/util.c] and [`src/sll/include/sll/util.h`][0.6.37/src/sll/include/sll/util.h] to [`src/sll/env.c`][0.6.38/src/sll/env.c] and [`src/sll/include/sll/env.h`][0.6.38/src/sll/include/sll/env.h]
 - Moved `sll_version` from [`src/sll/util.c`][0.6.37/src/sll/util.c] and [`src/sll/include/sll/util.h`][0.6.37/src/sll/include/sll/util.h] to [`src/sll/version.c`][0.6.38/src/sll/version.c] and [`src/sll/include/sll/version.h`][0.6.38/src/sll/include/sll/version.h]
 - Renamed `sll_api_time_current_nanos`, `sll_api_time_sleep_nanos`, `time$time_nanos` and `time$sleep_nanos` to `sll_api_time_current_ns`, `sll_api_time_sleep_ns`, `time$time_ns` and `time$sleep_ns`
 - Renamed [`src/sll/api/_generated.c`][0.6.37/src/sll/api/_generated.c] and [`src/sll/include/sll/api/_generated.h`][0.6.37/src/sll/include/sll/api/_generated.h] to [`src/sll/api/_generated_raw.c`][0.6.38/src/sll/api/_generated_raw.c] and [`src/sll/include/sll/api/_generated_raw.h`][0.6.38/src/sll/include/sll/api/_generated_raw.h]
@@ -1626,9 +1628,13 @@ Unfortunately, no versions were assigned before 2021-6-15 (:disappointed:), so t
 [test-coverage]: https://github.com/sl-lang/sll/tree/test-coverage
 [0.6.38/src/sll/api/_generated_raw.c]: https://github.com/sl-lang/sll/blob/main/src/sll/api/_generated_raw.c
 [0.6.38/src/sll/data.c]: https://github.com/sl-lang/sll/blob/main/src/sll/data.c
+[0.6.38/src/sll/env.c]: https://github.com/sl-lang/sll/blob/main/src/sll/env.c
 [0.6.38/src/sll/include/sll/api/_generated_raw.h]: https://github.com/sl-lang/sll/blob/main/src/sll/include/sll/api/_generated_raw.h
 [0.6.38/src/sll/include/sll/data.h]: https://github.com/sl-lang/sll/blob/main/src/sll/include/sll/data.h
+[0.6.38/src/sll/include/sll/env.h]: https://github.com/sl-lang/sll/blob/main/src/sll/include/sll/env.h
+[0.6.38/src/sll/include/sll/init.h]: https://github.com/sl-lang/sll/blob/main/src/sll/include/sll/init.h
 [0.6.38/src/sll/include/sll/version.h]: https://github.com/sl-lang/sll/blob/main/src/sll/include/sll/version.h
+[0.6.38/src/sll/init.c]: https://github.com/sl-lang/sll/blob/main/src/sll/init.c
 [0.6.38/src/sll/version.c]: https://github.com/sl-lang/sll/blob/main/src/sll/version.c
 [0.6.37/src/sll/api/_generated.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.37/src/sll/api/_generated.c
 [0.6.37/src/sll/include/sll/api/_generated.h]: https://github.com/sl-lang/sll/blob/sll-v0.6.37/src/sll/include/sll/api/_generated.h